research_assistant_agent/backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-core
langchain-community
python-dotenv
pypdf
duckduckgo-search
google-genai
